+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
+- Drafthorse (Python dependency) updated to version 2025.2
 - Weasyprint (Python dependency) updated to version 66
 - Schwifty (Python dependency) updated to version 2025.9
 - Remove brackets surrounding the date from the project name in Aktivenabrechnung e-invoices